Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    float a;
+    double float a;
     scanf("%f",&a);
     if(a==12345.6789){
         printf("12345.6789");
